+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for building OCI bundles
 - Support for LZ4 compression of builds
-- E2E cross-platform tests in GitHub Actions
 - **[BREAKING]** Expose `RIVET_API_ENDPOINT` to `cdn.build_command` to help automate deploying to multiple cluters
 - **[BREAKING]** Unset `RIVET_TOKEN` to `cdn.build_command` in order to ensure the cloud token isn't accidentally baked in a build
+- `image build-push` command to automatically build & push an image
+- `site build-push` command to automatially build and push a site
+- E2E cross-platform tests in GitHub Actions
 
 ### Changed